Fix type comparison issue from switching back to gtest 1.8.0

+++ tests/CBC_CS1_tests.cpp
@@ -58,7 +58,7 @@ TEST(CBC_CS1, KAT_enc)
 
 				res = ctx.finish(pad_sz);
 				EXPECT_EQ(res, 0);
-				EXPECT_EQ(pad_sz, 0);
+				EXPECT_EQ(pad_sz, (size_t)0);
 
 				output_sz = total_sz - current_sz;
 				res = ctx.steal_last(output.get() + current_sz, output_sz);
@@ -122,7 +122,7 @@ TEST(CBC_CS1, MMT_enc)
 
 				res = ctx.finish(pad_sz);
 				EXPECT_EQ(res, 0);
-				EXPECT_EQ(pad_sz, 0);
+				EXPECT_EQ(pad_sz, (size_t)0);
 
 				output_sz = total_sz - current_sz;
 				res = ctx.steal_last(output.get() + current_sz, output_sz);
@@ -187,7 +187,7 @@ TEST(CBC_CS1, MonteCarlo_enc)
 
 					res = ctx.finish(pad_sz);
 					EXPECT_EQ(res, 0);
-					EXPECT_EQ(pad_sz, 0);
+					EXPECT_EQ(pad_sz, (size_t)0);
 
 					output_sz = total_sz - current_sz;
 					res = ctx.steal_last(output[1] + current_sz, output_sz);
@@ -354,7 +354,7 @@ TEST(CBC_CS1, steal_test_vector_enc)
 
 		res = ctx.finish(pad_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(pad_sz, 0);
+		EXPECT_EQ(pad_sz, (size_t)0);
 
 		current_sz = total_sz - output_sz;
 		res = ctx.steal_last(output.get() + output_sz, current_sz);
@@ -415,7 +415,7 @@ TEST(CBC_CS1, KAT_dec)
 
 				res = ctx.finish(pad_sz);
 				EXPECT_EQ(res, 0);
-				EXPECT_EQ(pad_sz, 0);
+				EXPECT_EQ(pad_sz, (size_t)0);
 
 				output_sz = total_sz - current_sz;
 				res = ctx.steal_last(output.get() + current_sz, output_sz);
@@ -479,7 +479,7 @@ TEST(CBC_CS1, MMT_dec)
 
 				res = ctx.finish(pad_sz);
 				EXPECT_EQ(res, 0);
-				EXPECT_EQ(pad_sz, 0);
+				EXPECT_EQ(pad_sz, (size_t)0);
 
 				output_sz = total_sz - current_sz;
 				res = ctx.steal_last(output.get() + current_sz, output_sz);
@@ -544,7 +544,7 @@ TEST(CBC_CS1, MonteCarlo_dec)
 
 					res = ctx.finish(pad_sz);
 					EXPECT_EQ(res, 0);
-					EXPECT_EQ(pad_sz, 0);
+					EXPECT_EQ(pad_sz, (size_t)0);
 
 					output_sz = total_sz - current_sz;
 					res = ctx.steal_last(output[1] + current_sz, output_sz);
@@ -713,7 +713,7 @@ TEST(CBC_CS1, steal_test_vector_dec)
 
 		res = ctx.finish(pad_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(pad_sz, 0);
+		EXPECT_EQ(pad_sz, (size_t)0);
 
 		current_sz = total_sz - output_sz;
 		res = ctx.steal_last(output.get() + output_sz, current_sz);
@@ -751,7 +751,7 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 0, input_sz = 8, output_sz = 0
@@ -762,7 +762,7 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 0, input_sz = 24, output_sz = 0
@@ -773,7 +773,7 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 0, input_sz = 32, output_sz = 0
@@ -784,7 +784,7 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 1);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 0, input_sz = 32, output_sz = 16
@@ -795,7 +795,7 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 16;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 15, input_sz = 0, output_sz = 0
@@ -806,13 +806,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 0;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 15, input_sz = 1, output_sz = 0
@@ -823,13 +823,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 1;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 15, input_sz = 16, output_sz = 0
@@ -840,13 +840,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 16;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 15, input_sz = 17, output_sz = 0
@@ -857,13 +857,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 17;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 1);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 15, input_sz = 17, output_sz = 16
@@ -874,13 +874,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 17;
 		output_sz = 16;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 16, input_sz = 0, output_sz = 0
@@ -891,13 +891,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 0;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 16, input_sz = 15, output_sz = 0
@@ -908,13 +908,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 15;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 16, input_sz = 16, output_sz = 0
@@ -925,13 +925,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 16;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 1);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 16, input_sz = 16, output_sz = 16
@@ -942,13 +942,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 16;
 		output_sz = 16;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 16, input_sz = 24, output_sz = 16
@@ -959,13 +959,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 24;
 		output_sz = 16;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 31, input_sz = 0, output_sz = 0
@@ -976,13 +976,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 0;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 	}
 
 	// buffer_sz = 31, input_sz = 1, output_sz = 0
@@ -993,13 +993,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 1;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 1);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 31, input_sz = 1, output_sz = 16
@@ -1010,13 +1010,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 1;
 		output_sz = 16;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 31, input_sz = 16, output_sz = 16
@@ -1027,13 +1027,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 16;
 		output_sz = 16;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 31, input_sz = 33, output_sz = 0
@@ -1044,13 +1044,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 33;
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 1);
-		EXPECT_EQ(output_sz, 48);
+		EXPECT_EQ(output_sz, (size_t)48);
 	}
 
 	// buffer_sz = 31, input_sz = 33, output_sz = 48
@@ -1061,13 +1061,13 @@ TEST(CBC_CS1, update_sz)
 		output_sz = 0;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 0);
+		EXPECT_EQ(output_sz, (size_t)0);
 
 		input_sz = 33;
 		output_sz = 48;
 		res = ctx.update(input, input_sz, output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 48);
+		EXPECT_EQ(output_sz, (size_t)48);
 	}
 }
 
@@ -1095,7 +1095,7 @@ TEST(CBC_CS1, finish_sz)
 
 		res = ctx.finish(pad_sz);
 		EXPECT_EQ(res, 2);
-		EXPECT_EQ(pad_sz, 16);
+		EXPECT_EQ(pad_sz, (size_t)16);
 	}
 
 	// Buffer not empty < BLOCK_SIZE
@@ -1107,7 +1107,7 @@ TEST(CBC_CS1, finish_sz)
 
 		res = ctx.finish(pad_sz);
 		EXPECT_EQ(res, 2);
-		EXPECT_EQ(pad_sz, 1);
+		EXPECT_EQ(pad_sz, (size_t)1);
 	}
 
 	// Buffer not empty = BLOCK_SIZE
@@ -1119,7 +1119,7 @@ TEST(CBC_CS1, finish_sz)
 
 		res = ctx.finish(pad_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(pad_sz, 0);
+		EXPECT_EQ(pad_sz, (size_t)0);
 	}
 
 	// Buffer not empty > BLOCK_SIZE
@@ -1131,7 +1131,7 @@ TEST(CBC_CS1, finish_sz)
 
 		res = ctx.finish(pad_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(pad_sz, 0);
+		EXPECT_EQ(pad_sz, (size_t)0);
 	}
 }
 
@@ -1158,7 +1158,7 @@ TEST(CBC_CS1, steal_last_sz)
 		output_sz = 0;
 		res = ctx.steal_last(output, output_sz);
 		EXPECT_EQ(res, 2);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 15, output_sz = 0
@@ -1173,7 +1173,7 @@ TEST(CBC_CS1, steal_last_sz)
 		output_sz = 0;
 		res = ctx.steal_last(output, output_sz);
 		EXPECT_EQ(res, 2);
-		EXPECT_EQ(output_sz, 1);
+		EXPECT_EQ(output_sz, (size_t)1);
 	}
 
 	// buffer_sz = 16, output_sz = 0
@@ -1188,7 +1188,7 @@ TEST(CBC_CS1, steal_last_sz)
 		output_sz = 0;
 		res = ctx.steal_last(output, output_sz);
 		EXPECT_EQ(res, 1);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 16, output_sz = 16
@@ -1203,7 +1203,7 @@ TEST(CBC_CS1, steal_last_sz)
 		output_sz = 16;
 		res = ctx.steal_last(output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 16);
+		EXPECT_EQ(output_sz, (size_t)16);
 	}
 
 	// buffer_sz = 31, output_sz = 0
@@ -1218,7 +1218,7 @@ TEST(CBC_CS1, steal_last_sz)
 		output_sz = 0;
 		res = ctx.steal_last(output, output_sz);
 		EXPECT_EQ(res, 1);
-		EXPECT_EQ(output_sz, 31);
+		EXPECT_EQ(output_sz, (size_t)31);
 	}
 
 	// buffer_sz = 31, output_sz = 31
@@ -1233,6 +1233,6 @@ TEST(CBC_CS1, steal_last_sz)
 		output_sz = 31;
 		res = ctx.steal_last(output, output_sz);
 		EXPECT_EQ(res, 0);
-		EXPECT_EQ(output_sz, 31);
+		EXPECT_EQ(output_sz, (size_t)31);
 	}
 }
